e300: Type-cast fix for boost
Fixes compiler errors on newer Boost. boost::posix_time::microseconds()
requires a int64_t.

+++ host/lib/usrp/e300/e300_fifo_config.cpp
@@ -108,7 +108,7 @@ struct e300_fifo_poll_waiter
         boost::mutex::scoped_lock l(_mutex);
         if (_poll_claimed)
         {
-            _cond.timed_wait(l, boost::posix_time::microseconds(timeout*1000000));
+            _cond.timed_wait(l, boost::posix_time::microseconds(long(timeout*1000000)));
         }
         else
         {
